Add files via upload
dashboard creation using power bi
1. Import the Business data file into Microsoft PowerBI. 
2. Load only the “Account” sheet from the workbook after transforming the 
data. 
3. Represent the data as per the given criteria: 
a. Represent the data, which includes the total count of account numbers 
against the country, in a tabular format. 
b. Insert another table and represent the data on the country-wise Account 
holder's name. 
c. Insert a slicer from the visualisation section and add a country-wise filter. 
d. Represent the data of India, Brazil, Greece, and France together.